+++ extensions/web-clipper/src/api.ts
@@ -16,7 +16,7 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-import { browser, Runtime } from "webextension-polyfill-ts";
+import { browser, Runtime, Tabs } from "webextension-polyfill-ts";
 import { Remote, wrap } from "comlink";
 import { createEndpoint } from "./utils/comlink-extension";
 import { Server } from "./common/bridge";
@@ -28,14 +28,32 @@ let api: Remote<Server> | undefined;
 export async function connectApi(openNew = false, onDisconnect?: () => void) {
   if (api) return api;
 
-  const tab = await getTab(openNew);
-  if (!tab) return false;
+  const tabs = await findNotesnookTabs(openNew);
+  for (const tab of tabs) {
+    try {
+      const api = await Promise.race([
+        connectToTab(tab, onDisconnect),
+        timeout(5000)
+      ]);
+      if (!api) continue;
+      return api as Remote<Server>;
+    } catch (e) {
+      console.error(e);
+    }
+  }
 
-  return await new Promise<Remote<Server> | false>(function connect(resolve) {
+  return false;
+}
+
+function timeout(ms: number) {
+  return new Promise((resolve) => setTimeout(resolve, ms, false));
+}
+
+function connectToTab(tab: Tabs.Tab, onDisconnect?: () => void) {
+  return new Promise<Remote<Server> | false>(function connect(resolve) {
     if (!tab.id) return resolve(false);
 
     const port = browser.tabs.connect(tab.id);
-
     port.onDisconnect.addListener(() => {
       api = undefined;
       onDisconnect?.();
@@ -58,23 +76,26 @@ export async function connectApi(openNew = false, onDisconnect?: () => void) {
   });
 }
 
-async function getTab(openNew = false) {
+export async function findNotesnookTabs(openNew = false) {
   const tabs = await browser.tabs.query({
-    url: APP_URL_FILTER
+    url: APP_URL_FILTER,
+    discarded: false,
+    status: "complete"
   });
 
-  if (tabs.length) return tabs[0];
+  if (tabs.length) return tabs;
 
   if (openNew) {
-    const [tab] = await Promise.all([
-      browser.tabs.create({ url: APP_URL, active: false }),
-      new Promise((resolve) => {
-        browser.runtime.onMessage.addListener((message) => {
-          if (message.type === "start_connection") resolve(true);
-        });
+    const tab = await browser.tabs.create({ url: APP_URL, active: false });
+    await new Promise<void>((resolve) =>
+      browser.tabs.onUpdated.addListener(function onUpdated(id, info) {
+        if (id === tab.id && info.status === "complete") {
+          browser.tabs.onUpdated.removeListener(onUpdated);
+          resolve();
+        }
       })
-    ]);
-    return tab;
+    );
+    return [tab];
   }
-  return undefined;
+  return [];
 }
